+++ src/components/calcite-button/calcite-button.e2e.ts
@@ -8,7 +8,12 @@ describe("calcite-button", () => {
     const element = await page.find("calcite-button");
     const elementAsButton = await page.find("calcite-button >>> button");
     const elementAsLink = await page.find("calcite-button >>> a");
-    const icon = await page.find("calcite-button >>> .calcite-button--icon");
+    const iconStart = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-start"
+    );
+    const iconEnd = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-end"
+    );
     const loader = await page.find(
       "calcite-button >>> .calcite-button--loader"
     );
@@ -20,7 +25,8 @@ describe("calcite-button", () => {
     expect(element).toEqualAttribute("width", "auto");
     expect(elementAsLink).toBeNull();
     expect(elementAsButton).not.toBeNull();
-    expect(icon).toBeNull();
+    expect(iconStart).toBeNull();
+    expect(iconEnd).toBeNull();
     expect(loader).toBeNull();
   });
 
@@ -30,7 +36,12 @@ describe("calcite-button", () => {
     const element = await page.find("calcite-button");
     const elementAsButton = await page.find("calcite-button >>> button");
     const elementAsLink = await page.find("calcite-button >>> a");
-    const icon = await page.find("calcite-button >>> .calcite-button--icon");
+    const iconStart = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-start"
+    );
+    const iconEnd = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-end"
+    );
     const loader = await page.find(
       "calcite-button >>> .calcite-button--loader"
     );
@@ -42,7 +53,8 @@ describe("calcite-button", () => {
     expect(element).toEqualAttribute("width", "auto");
     expect(elementAsLink).not.toBeNull();
     expect(elementAsButton).toBeNull();
-    expect(icon).toBeNull();
+    expect(iconStart).toBeNull();
+    expect(iconEnd).toBeNull();
     expect(loader).toBeNull();
   });
 
@@ -54,7 +66,12 @@ describe("calcite-button", () => {
     const element = await page.find("calcite-button");
     const elementAsButton = await page.find("calcite-button >>> button");
     const elementAsLink = await page.find("calcite-button >>> a");
-    const icon = await page.find("calcite-button >>> .calcite-button--icon");
+    const iconStart = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-start"
+    );
+    const iconEnd = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-end"
+    );
     const loader = await page.find(
       "calcite-button >>> .calcite-button--loader"
     );
@@ -66,7 +83,8 @@ describe("calcite-button", () => {
     expect(element).toEqualAttribute("width", "half");
     expect(elementAsLink).toBeNull();
     expect(elementAsButton).not.toBeNull();
-    expect(icon).toBeNull();
+    expect(iconStart).toBeNull();
+    expect(iconEnd).toBeNull();
     expect(loader).toBeNull();
   });
 
@@ -78,7 +96,12 @@ describe("calcite-button", () => {
     const element = await page.find("calcite-button");
     const elementAsButton = await page.find("calcite-button >>> button");
     const elementAsLink = await page.find("calcite-button >>> a");
-    const icon = await page.find("calcite-button >>> .calcite-button--icon");
+    const iconStart = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-start"
+    );
+    const iconEnd = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-end"
+    );
     const loader = await page.find(
       "calcite-button >>> .calcite-button--loader"
     );
@@ -90,7 +113,8 @@ describe("calcite-button", () => {
     expect(element).toEqualAttribute("width", "half");
     expect(elementAsLink).not.toBeNull();
     expect(elementAsButton).toBeNull();
-    expect(icon).toBeNull();
+    expect(iconStart).toBeNull();
+    expect(iconEnd).toBeNull();
     expect(loader).toBeNull();
   });
 
@@ -102,7 +126,12 @@ describe("calcite-button", () => {
     const element = await page.find("calcite-button");
     const elementAsButton = await page.find("calcite-button >>> button");
     const elementAsLink = await page.find("calcite-button >>> a");
-    const icon = await page.find("calcite-button >>> .calcite-button--icon");
+    const iconStart = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-start"
+    );
+    const iconEnd = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-end"
+    );
     const loader = await page.find(
       "calcite-button >>> .calcite-button--loader"
     );
@@ -113,7 +142,8 @@ describe("calcite-button", () => {
     expect(elementAsLink).toEqualAttribute("href", "google.com");
     expect(elementAsLink).toEqualAttribute("rel", "noopener noreferrer");
     expect(elementAsLink).toEqualAttribute("target", "_blank");
-    expect(icon).toBeNull();
+    expect(iconStart).toBeNull();
+    expect(iconEnd).toBeNull();
     expect(loader).toBeNull();
   });
 
@@ -125,7 +155,12 @@ describe("calcite-button", () => {
     const element = await page.find("calcite-button");
     const elementAsButton = await page.find("calcite-button >>> button");
     const elementAsLink = await page.find("calcite-button >>> a");
-    const icon = await page.find("calcite-button >>> .calcite-button--icon");
+    const iconStart = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-start"
+    );
+    const iconEnd = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-end"
+    );
     const loader = await page.find(
       "calcite-button >>> .calcite-button--loader"
     );
@@ -135,7 +170,8 @@ describe("calcite-button", () => {
     expect(elementAsButton).toHaveClass("mycustomclass");
     expect(elementAsButton).toEqualAttribute("type", "reset");
     expect(elementAsButton).toEqualAttribute("name", "myname");
-    expect(icon).toBeNull();
+    expect(iconStart).toBeNull();
+    expect(iconEnd).toBeNull();
     expect(loader).toBeNull();
   });
 
@@ -147,7 +183,12 @@ describe("calcite-button", () => {
     const element = await page.find("calcite-button");
     const elementAsButton = await page.find("calcite-button >>> button");
     const elementAsLink = await page.find("calcite-button >>> a");
-    const icon = await page.find("calcite-button >>> .calcite-button--icon");
+    const iconStart = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-start"
+    );
+    const iconEnd = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-end"
+    );
     const loader = await page.find(
       "calcite-button >>> .calcite-button--loader"
     );
@@ -158,45 +199,158 @@ describe("calcite-button", () => {
     expect(element).toEqualAttribute("width", "auto");
     expect(elementAsLink).toBeNull();
     expect(elementAsButton).not.toBeNull();
-    expect(icon).toBeNull();
+    expect(iconStart).toBeNull();
+    expect(iconEnd).toBeNull();
     expect(loader).toBeNull();
   });
 
-  it("renders with an icon", async () => {
+  it("renders with an icon-start", async () => {
     const page = await newE2EPage();
     await page.setContent(
-      `<calcite-button icon='plus'>Continue</calcite-button>`
+      `<calcite-button icon-start='plus'>Continue</calcite-button>`
     );
     const element = await page.find("calcite-button");
     const elementAsButton = await page.find("calcite-button >>> button");
     const elementAsLink = await page.find("calcite-button >>> a");
-    const icon = await page.find("calcite-button >>> .calcite-button--icon");
+    const iconStart = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-start"
+    );
+    const iconEnd = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-end"
+    );
     const loader = await page.find(
       "calcite-button >>> .calcite-button--loader"
     );
     expect(element).toHaveClass("hydrated");
     expect(elementAsLink).toBeNull();
     expect(elementAsButton).not.toBeNull();
-    expect(icon).not.toBeNull();
+    expect(iconStart).not.toBeNull();
+    expect(iconEnd).toBeNull();
     expect(loader).toBeNull();
   });
 
-  it("renders with a loader and an icon element when both icon and loader are present", async () => {
+  it("renders with an icon-end", async () => {
     const page = await newE2EPage();
     await page.setContent(
-      `<calcite-button loading icon='plus'>Continue</calcite-button>`
+      `<calcite-button icon-end='plus'>Continue</calcite-button>`
     );
     const element = await page.find("calcite-button");
     const elementAsButton = await page.find("calcite-button >>> button");
     const elementAsLink = await page.find("calcite-button >>> a");
-    const icon = await page.find("calcite-button >>> .calcite-button--icon");
+    const iconStart = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-start"
+    );
+    const iconEnd = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-end"
+    );
     const loader = await page.find(
       "calcite-button >>> .calcite-button--loader"
     );
     expect(element).toHaveClass("hydrated");
     expect(elementAsLink).toBeNull();
     expect(elementAsButton).not.toBeNull();
-    expect(icon).not.toBeNull();
+    expect(iconStart).toBeNull();
+    expect(iconEnd).not.toBeNull();
+    expect(loader).toBeNull();
+  });
+
+  it("renders with an icon-start and icon-end", async () => {
+    const page = await newE2EPage();
+    await page.setContent(
+      `<calcite-button icon-start='plus' icon-end='plus'>Continue</calcite-button>`
+    );
+    const element = await page.find("calcite-button");
+    const elementAsButton = await page.find("calcite-button >>> button");
+    const elementAsLink = await page.find("calcite-button >>> a");
+    const iconStart = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-start"
+    );
+    const iconEnd = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-end"
+    );
+    const loader = await page.find(
+      "calcite-button >>> .calcite-button--loader"
+    );
+    expect(element).toHaveClass("hydrated");
+    expect(elementAsLink).toBeNull();
+    expect(elementAsButton).not.toBeNull();
+    expect(iconStart).not.toBeNull();
+    expect(iconEnd).not.toBeNull();
+    expect(loader).toBeNull();
+  });
+
+  it("renders with a loader and an icon-start when both icon-start and loader are requested", async () => {
+    const page = await newE2EPage();
+    await page.setContent(
+      `<calcite-button loading icon-start='plus'>Continue</calcite-button>`
+    );
+    const element = await page.find("calcite-button");
+    const elementAsButton = await page.find("calcite-button >>> button");
+    const elementAsLink = await page.find("calcite-button >>> a");
+    const iconStart = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-start"
+    );
+    const iconEnd = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-end"
+    );
+    const loader = await page.find(
+      "calcite-button >>> .calcite-button--loader"
+    );
+    expect(element).toHaveClass("hydrated");
+    expect(elementAsLink).toBeNull();
+    expect(elementAsButton).not.toBeNull();
+    expect(iconStart).not.toBeNull();
+    expect(iconEnd).toBeNull();
+    expect(loader).not.toBeNull();
+  });
+
+  it("renders with a loader and an icon-end when both icon-end and loader are requested", async () => {
+    const page = await newE2EPage();
+    await page.setContent(
+      `<calcite-button loading icon-end='plus'>Continue</calcite-button>`
+    );
+    const element = await page.find("calcite-button");
+    const elementAsButton = await page.find("calcite-button >>> button");
+    const elementAsLink = await page.find("calcite-button >>> a");
+    const iconStart = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-start"
+    );
+    const iconEnd = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-end"
+    );
+    const loader = await page.find(
+      "calcite-button >>> .calcite-button--loader"
+    );
+    expect(element).toHaveClass("hydrated");
+    expect(elementAsLink).toBeNull();
+    expect(elementAsButton).not.toBeNull();
+    expect(iconStart).toBeNull();
+    expect(iconEnd).not.toBeNull();
+    expect(loader).not.toBeNull();
+  });
+
+  it("renders with a loader and an icon-start and icon-end when all are requested", async () => {
+    const page = await newE2EPage();
+    await page.setContent(
+      `<calcite-button loading icon-start='plus' icon-end='plus'>Continue</calcite-button>`
+    );
+    const element = await page.find("calcite-button");
+    const elementAsButton = await page.find("calcite-button >>> button");
+    const elementAsLink = await page.find("calcite-button >>> a");
+    const iconStart = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-start"
+    );
+    const iconEnd = await page.find(
+      "calcite-button >>> .calcite-button--icon.icon-end"
+    );
+    const loader = await page.find(
+      "calcite-button >>> .calcite-button--loader"
+    );
+    expect(element).toHaveClass("hydrated");
+    expect(elementAsLink).toBeNull();
+    expect(elementAsButton).not.toBeNull();
+    expect(iconStart).not.toBeNull();
+    expect(iconEnd).not.toBeNull();
     expect(loader).not.toBeNull();
   });
 
@@ -209,7 +363,9 @@ describe("calcite-button", () => {
 
   it("hastext is false when text is not present", async () => {
     const page = await newE2EPage();
-    await page.setContent(`<calcite-button icon='plus'></calcite-button>`);
+    await page.setContent(
+      `<calcite-button icon-start='plus'></calcite-button>`
+    );
     const element = await page.find("calcite-button");
     expect(element).not.toHaveAttribute("hastext");
   });