+++ src/components/calcite-button/calcite-button.stories.js
@@ -3,7 +3,7 @@ import { withKnobs, text, boolean, select } from "@storybook/addon-knobs";
 import {
   darkBackground,
   iconNames,
-  parseReadme
+  parseReadme,
 } from "../../../.storybook/helpers";
 import readme from "./readme.md";
 const notes = parseReadme(readme);
@@ -33,7 +33,7 @@ storiesOf("Button", module)
     { notes }
   )
   .add(
-    "With icon",
+    "With icon-start",
     () => `
     <calcite-button
       appearance="${select(
@@ -48,8 +48,52 @@ storiesOf("Button", module)
       href="${text("href", "")}"
       loading="${boolean("loading", false)}"
       disabled="${boolean("disabled", false)}"
-      icon="${select("icon", iconNames, iconNames[0])}"
-      icon-position="${select("icon-position", ["start", "end"], "start")}">
+      icon-start="${select("icon-start", iconNames, iconNames[0])}">
+      ${text("text", "button text here")}
+    </calcite-button>
+  `,
+    { notes }
+  )
+  .add(
+    "With icon-end",
+    () => `
+    <calcite-button
+      appearance="${select(
+        "appearance",
+        ["solid", "clear", "outline", "transparent"],
+        "solid"
+      )}"
+      color="${select("color", ["blue", "red", "dark", "light"], "blue")}"
+      scale="${select("scale", ["s", "m", "l"], "m")}"
+      round="${boolean("round", false)}"
+      floating="${boolean("floating", false)}"
+      href="${text("href", "")}"
+      loading="${boolean("loading", false)}"
+      disabled="${boolean("disabled", false)}"
+      icon-end="${select("icon-end", iconNames, iconNames[0])}">
+      ${text("text", "button text here")}
+    </calcite-button>
+  `,
+    { notes }
+  )
+  .add(
+    "With icon-start and icon-end",
+    () => `
+    <calcite-button
+      appearance="${select(
+        "appearance",
+        ["solid", "clear", "outline", "transparent"],
+        "solid"
+      )}"
+      color="${select("color", ["blue", "red", "dark", "light"], "blue")}"
+      scale="${select("scale", ["s", "m", "l"], "m")}"
+      round="${boolean("round", false)}"
+      floating="${boolean("floating", false)}"
+      href="${text("href", "")}"
+      loading="${boolean("loading", false)}"
+      disabled="${boolean("disabled", false)}"
+      icon-start="${select("icon-start", iconNames, iconNames[0])}"
+      icon-end="${select("icon-end", iconNames, iconNames[0])}">
       ${text("text", "button text here")}
     </calcite-button>
   `,
@@ -61,8 +105,7 @@ storiesOf("Button", module)
     <div style="width: 480px; max-width: 100%; background-color: #fff">
       <calcite-button
         width="${select("width", ["auto", "half", "full"], "auto")}"
-        icon="${select("icon", iconNames, iconNames[0])}"
-        icon-position="${select("icon-position", ["start", "end"], "start")}">
+        icon-start="${select("icon-start", iconNames, iconNames[0])}">
         ${text("text", "button text here")}
       </calcite-button>
     </div>
@@ -91,7 +134,8 @@ storiesOf("Button", module)
       "solid"
     )}"
     color="${select("color-2", ["blue", "red", "dark", "light"], "blue")}"
-    icon="${select("icon", iconNames, iconNames[0])}">
+    icon-start="${select("icon-start", iconNames, iconNames[0])}"
+    >
     ${text("text-2", "Some long string")}
     </calcite-button>
   </div>
@@ -125,7 +169,7 @@ storiesOf("Button", module)
   round="${boolean("round", true)}"
   floating="${boolean("floating", true)}"
   width="${select("width", ["auto", "half", "full"], "auto")}"
-  icon="${select("icon", iconNames, iconNames[0])}"
+  icon-start="${select("icon-start", iconNames, iconNames[0])}"
   appearance="${select("appearance", ["solid", "outline"], "solid")}"
   color="${select("color", ["blue", "red", "dark", "light"], "blue")}"
   scale="${select("scale", ["s", "m", "l"], "m")}"></calcite-button>
@@ -162,7 +206,7 @@ storiesOf("Button", module)
     round="${boolean("round", true)}"
     floating="${boolean("floating", true)}"
     width="${select("width", ["auto", "half", "full"], "auto")}"
-    icon="${select("icon", iconNames, iconNames[0])}"
+    icon-start="${select("icon-start", iconNames, iconNames[0])}"
     appearance="${select("appearance", ["solid", "outline"], "solid")}"
     color="${select("color", ["blue", "red", "dark", "light"], "blue")}"
     scale="${select("scale", ["s", "m", "l"], "m")}">
@@ -190,7 +234,9 @@ storiesOf("Button", module)
     href="${text("href", "")}"
     loading="${boolean("loading", false)}"
     disabled="${boolean("disabled", false)}"
-    icon="${select("icon", iconNames, iconNames[0])}">
+    icon-start="${select("icon-start", iconNames, iconNames[0])}"
+    icon-end="${select("icon-end", iconNames, iconNames[0])}"
+    >
     ${text("text", "button text here")}
   </calcite-button>
   `,
@@ -224,7 +270,7 @@ storiesOf("Button", module)
     round="${boolean("round", true)}"
     floating="${boolean("floating", true)}"
     width="${select("width", ["auto", "half", "full"], "auto")}"
-    icon="${select("icon", iconNames, iconNames[0])}"
+    icon-start="${select("icon-start", iconNames, iconNames[0])}"
     appearance="${select("appearance", ["solid", "outline"], "solid")}"
     color="${select("color", ["blue", "red", "dark", "light"], "blue")}"
     scale="${select("scale", ["s", "m", "l"], "m")}"></calcite-button>
@@ -261,7 +307,7 @@ storiesOf("Button", module)
   round="${boolean("round", true)}"
   floating="${boolean("floating", true)}"
   width="${select("width", ["auto", "half", "full"], "auto")}"
-  icon="${select("icon", iconNames, iconNames[0])}"
+  icon-start="${select("icon-start", iconNames, iconNames[0])}"
   appearance="${select("appearance", ["solid", "outline"], "solid")}"
   color="${select("color", ["blue", "red", "dark", "light"], "blue")}"
   scale="${select("scale", ["s", "m", "l"], "m")}">